+++ src/main/java/com/google/edith/servlets/DealsServlet.java
@@ -12,42 +12,17 @@ import javax.servlet.http.HttpServletResponse;
 @WebServlet("/deals")
 public class DealsServlet extends HttpServlet {
 
-  private final DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
   private static final String apiKey = "BKFuTzor";
   private static final String apiSite = "https://api.discount.com/v2/deals?api-key=%s";
-  private static Map<String, Double> receiptMap = new HashMap<String, Double>();
 
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    sumReceipt();
-    storeReceipt();
-    getDeals();
+
   }
 
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    String itemName = String.parseString(request.getParameter("itemName"));
-    Double itemPrice = Double.parseDouble(request.getParameter("itemPrice"));
-    receiptMap.put(itemName, itemPrice);
-  }
 
-  // Store a receipt in Datastore.
-  private void storeReceipt() {
-    Entity receiptEntity = new Entity("Receipt");
-    receiptEntity.setProperty()
-    for (String item: receiptMap.keySet()) {
-      //also pass receipt key when we figure out how to generate
-      storeItem(item);
-    }
-    datastore.put(receiptEntity);
-  }
-
-  // Store a grocery item in Datastore.
-  private void storeItem(String item) {
-    Entity itemEntity = new Entity("Item");
-    itemEntity.setProperty("name", item);
-    itemEntity.setProperty("price", receiptMap.get(item));
-    datastore.put(itemEntity);
   }
 
   // Fetch deals from Discount API.
@@ -76,8 +51,4 @@ public class DealsServlet extends HttpServlet {
     String json = gson.toJson(content);
   }
 
-  // Calculate total spending on shopping trip.
-  private void sumReceipt() {
-    
-  }
 }
